refactor: move bar ui classes into ui package
